requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
mysql-connector-python>=8.3.0
sqlalchemy>=2.0.0
joblib>=1.4.0
matplotlib>=3.8.0